refactor(hush-cli): extract test_cmd from pkg_cli
Move the `pkg test` cluster into pkg_cli/test_cmd.rs, carrying both
feature-gated cmd_pkg_test definitions verbatim (the
#[cfg(feature = "wasm-plugin-runtime")] runner and the #[cfg(not(...))]
stub) plus the wasm-gated helpers load_plugin_manifest_from_package,
find_wasm_binary, load_runtime_options. cmd_pkg_test widened to
pub(super) on both variants; helpers stay private. test_cmd's
sanitize_cargo_package_name / PLUGIN_MANIFEST_FILENAME / PathBuf imports
carry #[cfg(feature = "wasm-plugin-runtime")] to match. mod.rs imports
cmd_pkg_test unconditionally and drops the now-unused
sanitize_cargo_package_name import. Verified building with and without
the wasm-plugin-runtime feature.
